Guard rake task against net errors

+++ lib/tasks/nondisposable_tasks.rake
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'open-uri'
+require 'net/http'
 
 namespace :nondisposable do
   desc "Update the list of disposable email domains"
@@ -8,31 +9,34 @@ namespace :nondisposable do
     Rails.logger.info "Refreshing list of disposable domains..."
 
     url = 'https://raw.githubusercontent.com/disposable-email-domains/disposable-email-domains/master/disposable_email_blocklist.conf'
-    # url = 'http://localhost:3002/disposable_email_blocklist.conf'
 
     begin
-      downloaded_domains = URI.open(url).read.split("\n")
-      raise "The list is empty. This might indicate a problem with the format." if downloaded_domains.empty?
+      uri = URI(url)
+      response = Net::HTTP.get_response(uri)
 
-      Rails.logger.info "Downloaded list of disposable domains..."
+      if response.is_a?(Net::HTTPSuccess)
+        downloaded_domains = response.body.split("\n")
+        raise "The list is empty. This might indicate a problem with the format." if downloaded_domains.empty?
 
-      domains = (downloaded_domains + Nondisposable.configuration.additional_domains).uniq
-      domains -= Nondisposable.configuration.excluded_domains
+        Rails.logger.info "Downloaded list of disposable domains..."
 
-      ActiveRecord::Base.transaction do
-        Rails.logger.info "Updating disposable domains..."
-        Nondisposable::DisposableDomain.delete_all
+        domains = (downloaded_domains + Nondisposable.configuration.additional_domains).uniq
+        domains -= Nondisposable.configuration.excluded_domains
 
-        domains.map { |domain| Nondisposable::DisposableDomain.create(name: domain.downcase) }
+        ActiveRecord::Base.transaction do
+          Rails.logger.info "Updating disposable domains..."
+          Nondisposable::DisposableDomain.delete_all
+
+          domains.each { |domain| Nondisposable::DisposableDomain.create(name: domain.downcase) }
+        end
+
+        Rails.logger.info "Finished updating disposable domains. Total domains: #{domains.count}"
+      else
+        Rails.logger.error "Failed to download the list. HTTP Status: #{response.code}"
       end
-
-      Rails.logger.info "Refreshing cache..."
-      Nondisposable::DisposableDomain.refresh_cache
-
-      Rails.logger.info "Finished updating disposable domains."
-    rescue OpenURI::HTTPError => e
-      Rails.logger.error "An error occurred when trying to download the list of disposable domains: #{e.message}"
-    rescue => e
+    rescue SocketError => e
+      Rails.logger.error "Network error occurred: #{e.message}"
+    rescue StandardError => e
       Rails.logger.error "An error occurred when trying to update the list of disposable domains: #{e.message}"
     end
   end
